Fixed pendulum problem
