+++ ctsb/problems/tests/run_all_tests.py
@@ -6,6 +6,7 @@ from ctsb.problems.tests.test_rnn_output import test_rnn
 from ctsb.problems.tests.test_sp500 import test_sp500
 from ctsb.problems.tests.test_uci_indoor import test_uci_indoor
 from ctsb.problems.tests.test_crypto import test_crypto
+from ctsb.problems.tests.test_pendulum import test_pendulum
 
 def run_all_tests(show=True):
     print("\nrunning all problems tests...\n")
@@ -17,6 +18,7 @@ def run_all_tests(show=True):
     test_sp500(show_plot=show)
     test_uci_indoor(show_plot=show)
     test_crypto(show_plot=show)
+    test_pendulum(show_plot=show)
     print("\nall problems tests passed\n")
   
 if __name__ == "__main__":
